[Utils] Refactor utility functions in separate files (#49)
