+++ lib/Conversion/LowerDIP/LowerDIPPass.cpp
@@ -29,6 +29,8 @@
 
 #include "DIP/DIPDialect.h"
 #include "DIP/DIPOps.h"
+#include "Utils/DIPUtils.h"
+#include "Utils/Utils.h"
 #include <vector>
 
 using namespace mlir;
@@ -41,81 +43,6 @@ using namespace mlir::arith;
 //===----------------------------------------------------------------------===//
 
 namespace {
-// Calculate result of FMA and store it in output memref. This function cannot
-// handle tail processing.
-void calcAndStoreFMAwoTailProcessing(OpBuilder &builder, Location loc,
-                                     VectorType vecType, Value inputVec,
-                                     Value kernelVec, Value output,
-                                     Value beginIdx, Value endIdx) {
-  Value outputVec = builder.create<LoadOp>(loc, vecType, output,
-                                           ValueRange{beginIdx, endIdx});
-  Value resVec = builder.create<FMAOp>(loc, inputVec, kernelVec, outputVec);
-  builder.create<StoreOp>(loc, resVec, output, ValueRange{beginIdx, endIdx});
-}
-
-Value tailChecker(OpBuilder &builder, Location loc, AffineMap calcHelper,
-                  Value strideVal, Value kernelSize, Value c1, Value pseudoCol,
-                  Value colPivot) {
-  Value tailChecker = builder.create<AffineApplyOp>(
-      loc, calcHelper, ValueRange{strideVal, kernelSize, c1});
-  Value colEndDistance = builder.create<SubIOp>(loc, pseudoCol, colPivot);
-  Value tailCond = builder.create<CmpIOp>(loc, CmpIPredicate::sge,
-                                          colEndDistance, tailChecker);
-  return tailCond;
-}
-
-Value tailMaskCreator(OpBuilder &builder, Location loc, Value inputCol,
-                      Value colPivot, VectorType vectorMaskTy) {
-  Value extraElemCount = builder.create<SubIOp>(loc, inputCol, colPivot);
-  Value tailMask =
-      builder.create<CreateMaskOp>(loc, vectorMaskTy, extraElemCount);
-  return tailMask;
-}
-
-// Calculate result of FMA and store it in output memref. This function can
-// handle tail processing.
-void calcAndStoreFMAwTailProcessing(OpBuilder &builder, Location loc,
-                                    VectorType vecType, Value inputVec,
-                                    Value kernelVec, Value output,
-                                    Value beginIdx, Value endIdx,
-                                    Value tailCond, Value zeroPadding,
-                                    Value inputCol, VectorType vectorMaskTy) {
-  builder.create<scf::IfOp>(
-      loc, tailCond,
-      [&](OpBuilder &builder, Location loc) {
-        Value outputVec = builder.create<LoadOp>(loc, vecType, output,
-                                                 ValueRange{beginIdx, endIdx});
-        Value resVec =
-            builder.create<FMAOp>(loc, inputVec, kernelVec, outputVec);
-        builder.create<StoreOp>(loc, resVec, output,
-                                ValueRange{beginIdx, endIdx});
-
-        builder.create<scf::YieldOp>(loc);
-      },
-      [&](OpBuilder &builder, Location loc) {
-        Value extraElemMask =
-            tailMaskCreator(builder, loc, inputCol, endIdx, vectorMaskTy);
-        Value outputVec = builder.create<MaskedLoadOp>(
-            loc, vecType, output, ValueRange{beginIdx, endIdx}, extraElemMask,
-            zeroPadding);
-        Value resVec =
-            builder.create<FMAOp>(loc, inputVec, kernelVec, outputVec);
-        builder.create<MaskedStoreOp>(loc, output, ValueRange{beginIdx, endIdx},
-                                      extraElemMask, resVec);
-
-        builder.create<scf::YieldOp>(loc);
-      });
-}
-
-// Create an inverted mask having all 1's shifted to right side.
-Value createInvertedMask(OpBuilder &builder, Location loc, Value strideVal,
-                         VectorType vectorMaskTy, Value leftIndex) {
-  Value leftMask = builder.create<CreateMaskOp>(loc, vectorMaskTy, leftIndex);
-  Value maskInverter =
-      builder.create<CreateMaskOp>(loc, vectorMaskTy, strideVal);
-  Value rightMask = builder.create<SubIOp>(loc, maskInverter, leftMask);
-  return rightMask;
-}
 
 class DIPCorr2DLowering : public OpRewritePattern<dip::Corr2DOp> {
 public:
@@ -571,274 +498,6 @@ public:
 private:
   int64_t stride;
 };
-
-// Cast a value from index type to f32 type.
-Value indexToF32(OpBuilder &builder, Location loc, Value val) {
-  Value interm1 =
-      builder.create<arith::IndexCastOp>(loc, builder.getI32Type(), val);
-  return builder.create<arith::SIToFPOp>(loc, builder.getF32Type(), interm1);
-}
-
-// Cast a value from f32 type to index type.
-Value F32ToIndex(OpBuilder &builder, Location loc, Value val) {
-  Value interm1 =
-      builder.create<arith::FPToUIOp>(loc, builder.getI32Type(), val);
-  return builder.create<arith::IndexCastOp>(loc, builder.getIndexType(),
-                                            interm1);
-}
-
-// Round off floating point value to nearest integer type value.
-Value roundOff(OpBuilder &builder, Location loc, Value val) {
-  Value ceilVal = builder.create<math::CeilOp>(loc, val);
-  Value floorVal = builder.create<math::FloorOp>(loc, val);
-
-  Value diffCeil = builder.create<arith::SubFOp>(loc, ceilVal, val);
-  Value diffFloor = builder.create<arith::SubFOp>(loc, val, floorVal);
-
-  Value diffCond = builder.create<arith::CmpFOp>(loc, CmpFPredicate::OGT,
-                                                 diffCeil, diffFloor);
-
-  return builder.create<arith::SelectOp>(loc, diffCond, floorVal, ceilVal);
-}
-
-// Bound values to permissible range of allocatable values w.r.t output image.
-Value valBound(OpBuilder &builder, Location loc, Value val, Value lastElemF32,
-               Value c0F32) {
-  Value interm1 = builder.create<arith::MaxFOp>(loc, val, c0F32);
-  return builder.create<arith::MinFOp>(loc, interm1, lastElemF32);
-}
-
-// Apply 3 shear method and return mapped values.
-std::vector<Value> shearTransform(OpBuilder &builder, Location loc,
-                                  Value originalX, Value originalY,
-                                  Value sinVec, Value tanVec) {
-  Value yTan1 = builder.create<arith::MulFOp>(loc, tanVec, originalY);
-  Value xIntermediate1 = builder.create<arith::SubFOp>(loc, originalX, yTan1);
-  Value xIntermediate = roundOff(builder, loc, xIntermediate1);
-
-  Value xSin = builder.create<arith::MulFOp>(loc, xIntermediate, sinVec);
-  Value newY1 = builder.create<arith::AddFOp>(loc, xSin, originalY);
-  Value newY = roundOff(builder, loc, newY1);
-
-  Value yTan2 = builder.create<arith::MulFOp>(loc, newY, tanVec);
-  Value newX1 = builder.create<arith::SubFOp>(loc, xIntermediate, yTan2);
-  Value newX = roundOff(builder, loc, newX1);
-
-  return {newY, newX};
-}
-
-// Apply standard rotation matrix transformation and return mapped values.
-std::vector<Value> standardRotate(OpBuilder &builder, Location loc,
-                                  Value originalX, Value originalY,
-                                  Value sinVec, Value cosVec) {
-  Value ySin = builder.create<arith::MulFOp>(loc, originalY, sinVec);
-  Value yCos = builder.create<arith::MulFOp>(loc, originalY, cosVec);
-
-  Value xSin = builder.create<arith::MulFOp>(loc, originalX, sinVec);
-  Value xCos = builder.create<arith::MulFOp>(loc, originalX, cosVec);
-
-  Value newY1 = builder.create<arith::SubFOp>(loc, yCos, xSin);
-  Value newX1 = builder.create<arith::AddFOp>(loc, ySin, xCos);
-
-  return {roundOff(builder, loc, newY1), roundOff(builder, loc, newX1)};
-}
-
-// Get center co-ordinates w.r.t given dimension.
-Value getCenter(OpBuilder &builder, Location loc, MLIRContext *ctx, Value dim) {
-  Value dimF32 = indexToF32(builder, loc, dim);
-  Value c1f = builder.create<ConstantFloatOp>(loc, (llvm::APFloat)1.0f,
-                                              builder.getF32Type());
-  Value c2f = builder.create<ConstantFloatOp>(loc, (llvm::APFloat)2.0f,
-                                              builder.getF32Type());
-
-  Value temp1 = builder.create<arith::AddFOp>(loc, dimF32, c1f);
-  Value temp2 = builder.create<arith::DivFOp>(loc, temp1, c2f);
-  Value center = builder.create<arith::SubFOp>(loc, temp2, c1f);
-  Value centerRound = roundOff(builder, loc, center);
-
-  return F32ToIndex(builder, loc, centerRound);
-}
-
-// Equivalent of std::iota.
-Value iotaVec(OpBuilder &builder, Location loc, MLIRContext *ctx,
-              Value indexStart, Value strideVal, VectorType vecType, Value c0,
-              int64_t stride) {
-  // ToDo : Try to get rid of memref load/store, find less expensive ways of
-  // implementing this function.
-  MemRefType memTy = MemRefType::get({stride}, builder.getF32Type());
-  Value tempMem = builder.create<memref::AllocOp>(loc, memTy);
-
-  builder.create<AffineForOp>(
-      loc, ValueRange{c0}, builder.getDimIdentityMap(), ValueRange{strideVal},
-      builder.getDimIdentityMap(), 1, llvm::None,
-      [&](OpBuilder &builder, Location loc, Value iv, ValueRange iterArg) {
-        Value iotaValIndex = builder.create<arith::AddIOp>(loc, iv, indexStart);
-        Value iotaVal = indexToF32(builder, loc, iotaValIndex);
-
-        builder.create<memref::StoreOp>(loc, iotaVal, tempMem, ValueRange{iv});
-        builder.create<AffineYieldOp>(loc);
-      });
-
-  return builder.create<vector::LoadOp>(loc, vecType, tempMem, ValueRange{c0});
-}
-
-// Scale pixel co-ordinates appropriately before calculating their rotated
-// position(s).
-Value pixelScaling(OpBuilder &builder, Location loc, Value imageDImF32Vec,
-                   Value coordVec, Value imageCenterF32Vec, Value c1F32Vec) {
-  Value interm1 = builder.create<arith::SubFOp>(loc, imageDImF32Vec, coordVec);
-  Value interm2 =
-      builder.create<arith::SubFOp>(loc, interm1, imageCenterF32Vec);
-
-  return builder.create<arith::SubFOp>(loc, interm2, c1F32Vec);
-}
-
-// Fill appropriate pixel data in its corresponding rotated co-ordinate of
-// output image.
-void fillPixels(OpBuilder &builder, Location loc, Value resXVec, Value resYVec,
-                Value xVec, Value yVec, Value input, Value output, Value c0,
-                Value strideVal, Value outputRowLastElemF32,
-                Value outputColLastElemF32, Value c0F32) {
-  builder.create<AffineForOp>(
-      loc, ValueRange{c0}, builder.getDimIdentityMap(), ValueRange{strideVal},
-      builder.getDimIdentityMap(), /*step*/ 1, llvm::None,
-      [&](OpBuilder &builder, Location loc, ValueRange ivs,
-          ValueRange iterArg) {
-        Value resXPos =
-            builder.create<vector::ExtractElementOp>(loc, resXVec, ivs[0]);
-        Value resYPos =
-            builder.create<vector::ExtractElementOp>(loc, resYVec, ivs[0]);
-
-        Value resXPosBound =
-            valBound(builder, loc, resXPos, outputColLastElemF32, c0F32);
-        Value resYPosBound =
-            valBound(builder, loc, resYPos, outputRowLastElemF32, c0F32);
-
-        Value resXPosIndex = F32ToIndex(builder, loc, resXPosBound);
-        Value resYPosIndex = F32ToIndex(builder, loc, resYPosBound);
-
-        Value xPos =
-            builder.create<vector::ExtractElementOp>(loc, xVec, ivs[0]);
-        Value yPos =
-            builder.create<vector::ExtractElementOp>(loc, yVec, ivs[0]);
-
-        Value xPosIndex = F32ToIndex(builder, loc, xPos);
-        Value yPosIndex = F32ToIndex(builder, loc, yPos);
-
-        Value pixelVal = builder.create<memref::LoadOp>(
-            loc, builder.getF32Type(), input, ValueRange{xPosIndex, yPosIndex});
-        builder.create<memref::StoreOp>(loc, pixelVal, output,
-                                        ValueRange{resXPosIndex, resYPosIndex});
-
-        builder.create<AffineYieldOp>(loc);
-      });
-}
-
-// Cast index type value to f32 type and then expand it in a vector.
-Value castAndExpand(OpBuilder &builder, Location loc, Value val,
-                    VectorType vecType) {
-  Value interm1 = indexToF32(builder, loc, val);
-  return builder.create<vector::SplatOp>(loc, vecType, interm1);
-}
-
-// Calculate tan(angle / 2) where angle is a function parameter.
-Value customTanVal(OpBuilder &builder, Location loc, Value angleVal) {
-  Value c2F32 = builder.create<ConstantFloatOp>(loc, (llvm::APFloat)2.0f,
-                                                builder.getF32Type());
-  Value angleVal_2 = builder.create<arith::DivFOp>(loc, angleVal, c2F32);
-
-  Value sinVal = builder.create<math::SinOp>(loc, angleVal_2);
-  Value cosVal = builder.create<math::CosOp>(loc, angleVal_2);
-
-  return builder.create<arith::DivFOp>(loc, sinVal, cosVal);
-}
-
-// Controls shear transform application.
-void shearTransformController(
-    OpBuilder &builder, Location loc, MLIRContext *ctx,
-    SmallVector<Value, 8> lowerBounds, SmallVector<Value, 8> upperBounds,
-    SmallVector<int64_t, 8> steps, Value strideVal, Value input, Value output,
-    Value sinVec, Value tanVec, Value inputRowF32Vec, Value inputColF32Vec,
-    Value inputCenterYF32Vec, Value inputCenterXF32Vec,
-    Value outputCenterYF32Vec, Value outputCenterXF32Vec,
-    Value outputRowLastElemF32, Value outputColLastElemF32, Value c0,
-    Value c0F32, Value c1F32Vec, VectorType vectorTy32, int64_t stride,
-    FloatType f32) {
-  buildAffineLoopNest(
-      builder, loc, lowerBounds, upperBounds, steps,
-      [&](OpBuilder &builder, Location loc, ValueRange ivs) {
-        Value xLowerBoundMult =
-            builder.create<arith::DivUIOp>(loc, ivs[1], strideVal);
-        Value xLowerBound =
-            builder.create<arith::MulIOp>(loc, strideVal, xLowerBoundMult);
-
-        Value ivs0F32 = indexToF32(builder, loc, ivs[0]);
-        Value yVec = builder.create<vector::SplatOp>(loc, vectorTy32, ivs0F32);
-        Value xVec = iotaVec(builder, loc, ctx, xLowerBound, strideVal,
-                             vectorTy32, c0, stride);
-
-        Value yVecModified = pixelScaling(builder, loc, inputRowF32Vec, yVec,
-                                          inputCenterYF32Vec, c1F32Vec);
-        Value xVecModified = pixelScaling(builder, loc, inputColF32Vec, xVec,
-                                          inputCenterXF32Vec, c1F32Vec);
-
-        std::vector<Value> resIndices = shearTransform(
-            builder, loc, xVecModified, yVecModified, sinVec, tanVec);
-        Value resYVec = builder.create<arith::SubFOp>(loc, outputCenterYF32Vec,
-                                                      resIndices[0]);
-        Value resXVec = builder.create<arith::SubFOp>(loc, outputCenterXF32Vec,
-                                                      resIndices[1]);
-
-        fillPixels(builder, loc, resXVec, resYVec, xVec, yVec, input, output,
-                   c0, strideVal, outputRowLastElemF32, outputColLastElemF32,
-                   c0F32);
-      });
-}
-
-// Controls standard rotation matrix application.
-void standardRotateController(
-    OpBuilder &builder, Location loc, MLIRContext *ctx,
-    SmallVector<Value, 8> lowerBounds, SmallVector<Value, 8> upperBounds,
-    SmallVector<int64_t, 8> steps, Value strideVal, Value input, Value output,
-    Value sinVec, Value angleVal, Value inputRowF32Vec, Value inputColF32Vec,
-    Value inputCenterYF32Vec, Value inputCenterXF32Vec,
-    Value outputCenterYF32Vec, Value outputCenterXF32Vec,
-    Value outputRowLastElemF32, Value outputColLastElemF32, Value c0,
-    Value c0F32, Value c1F32Vec, VectorType vectorTy32, int64_t stride,
-    FloatType f32) {
-  Value cosVal = builder.create<math::CosOp>(loc, angleVal);
-  Value cosVec = builder.create<vector::BroadcastOp>(loc, vectorTy32, cosVal);
-
-  buildAffineLoopNest(
-      builder, loc, lowerBounds, upperBounds, steps,
-      [&](OpBuilder &builder, Location loc, ValueRange ivs) {
-        Value xLowerBoundMult =
-            builder.create<arith::DivUIOp>(loc, ivs[1], strideVal);
-        Value xLowerBound =
-            builder.create<arith::MulIOp>(loc, strideVal, xLowerBoundMult);
-
-        Value ivs0F32 = indexToF32(builder, loc, ivs[0]);
-        Value yVec = builder.create<vector::SplatOp>(loc, vectorTy32, ivs0F32);
-        Value xVec = iotaVec(builder, loc, ctx, xLowerBound, strideVal,
-                             vectorTy32, c0, stride);
-
-        Value yVecModified = pixelScaling(builder, loc, inputRowF32Vec, yVec,
-                                          inputCenterYF32Vec, c1F32Vec);
-        Value xVecModified = pixelScaling(builder, loc, inputColF32Vec, xVec,
-                                          inputCenterXF32Vec, c1F32Vec);
-
-        std::vector<Value> resIndices = standardRotate(
-            builder, loc, xVecModified, yVecModified, sinVec, cosVec);
-        Value resYVec = builder.create<arith::SubFOp>(loc, outputCenterYF32Vec,
-                                                      resIndices[0]);
-        Value resXVec = builder.create<arith::SubFOp>(loc, outputCenterXF32Vec,
-                                                      resIndices[1]);
-
-        fillPixels(builder, loc, resXVec, resYVec, xVec, yVec, input, output,
-                   c0, strideVal, outputRowLastElemF32, outputColLastElemF32,
-                   c0F32);
-      });
-}
 
 class DIPRotate2DOpLowering : public OpRewritePattern<dip::Rotate2DOp> {
 public:
